requirements.txt
appnope==0.1.2
asgiref==3.4.1
asttokens==2.0.5
autopep8==1.5.7
backcall==0.2.0
backports.zoneinfo==0.2.1
black==22.1.0
certifi==2021.10.8
chardet==3.0.4
click==8.0.3
debugpy==1.5.1
decorator==5.1.1
dj-database-url==0.5.0
Django==2.2.21
django-geojson==2.12.0
django-leaflet==0.26.0
entrypoints==0.4
executing==0.8.2
gunicorn==20.1.0
idna==2.6
ipykernel==6.9.0
ipython==8.0.1
jedi==0.18.1
jsonfield==2.0.2
jupyter-client==7.1.2
jupyter-core==4.9.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
nest-asyncio==1.5.4
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.1
platformdirs==2.4.1
prompt-toolkit==3.0.27
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.7.0
PyGithub==1.35
Pygments==2.11.2
PyJWT==2.3.0
python-dateutil==2.8.2
python-decouple==3.1
pytz==2021.3
pyzmq==22.3.0
requests==2.18.4
six==1.16.0
sqlparse==0.4.2
stack-data==0.1.4
toml==0.10.2
tomli==2.0.1
tornado==6.1
traitlets==5.1.1
typing_extensions==4.0.1
Unipath==1.1
urllib3==1.22
wcwidth==0.2.5
whitenoise==5.3.0
